Change server log from "connecting online" to "connection online"
The client already says "got accept. connection online"
and "connecting online" sounds a bit weird. I asked some natives
and they told me "connection established" is the most badass way
of saying it. But I rather keep the change as minimal as possible.

This is a breaking change for users parsing the logs.