Create methods for the startinfo response
The client sending the startinfo is a crucial part of building up
a healthy connection.

The server responds with 3 or more (depending on the amount of votes) messages.
That used to be one big wall of code that required comments to
understand where a new message starts. Now the protocol is easier to
understand while looking at the code. Without the need of comments.

```C++
if (MsgID == NETMSGTYPE_CL_STARTINFO)
{

	// [..]

	SendVoteClearOptions(ClientID);
	SendVoteOptions(ClientID);
	SendTuningParams(ClientID);
	SendReadyToEnter(pPlayer);
}
```